disaster/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-extensions==3.2.3
djangorestframework==3.14.0
idna==3.4
numpy==1.25.1
pandas==2.0.3
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rosetta==0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
django-plotly-dash==2.2.1
django-bootstrap-v5==1.0.11